+++ kitty/kitty_bridge.c
@@ -239,7 +239,12 @@ static INT_PTR CALLBACK KittyAboutProc(HWND hwnd, UINT msg,
     char buffer[1024];
     switch (msg) {
       case WM_INITDIALOG:
+#ifdef KITTY_TEST_BUILD_LABEL
+        sprintf(buffer, "KiTTY - %s\r\nTEST BUILD: %s", BuildVersionTime,
+                KITTY_TEST_BUILD_LABEL);
+#else
         sprintf(buffer, "KiTTY - %s", BuildVersionTime);
+#endif
         SetDlgItemText(hwnd, IDA_VERSION, buffer);
         return 1;
       case WM_COMMAND: